+++ src-opam/dockerfile_distro.ml
@@ -117,7 +117,7 @@ type t = [
   | `Archlinux of [ `Latest ]
   | `CentOS of [ `V6 | `V7 | `V8 | `Latest ]
   | `Debian of [ `V11 | `V10 | `V9 | `V8 | `V7 | `Stable | `Testing | `Unstable ]
-  | `Fedora of [ `V21 | `V22 | `V23 | `V24 | `V25 | `V26 | `V27 | `V28 | `V29 | `V30 | `V31 | `V32 | `V33 | `V34 | `Latest ]
+  | `Fedora of [ `V21 | `V22 | `V23 | `V24 | `V25 | `V26 | `V27 | `V28 | `V29 | `V30 | `V31 | `V32 | `V33 | `V34 | `V35 | `Latest ]
   | `OracleLinux of [ `V7 | `V8 | `Latest ]
   | `OpenSUSE of [ `V42_1 | `V42_2 | `V42_3 | `V15_0 | `V15_1 | `V15_2 | `V15_3 | `Latest ]
   | `Ubuntu of [ `V12_04 | `V14_04 | `V15_04 | `V15_10 | `V16_04 | `V16_10 | `V17_04 | `V17_10 | `V18_04 | `V18_10 | `V19_04 | `V19_10 | `V20_04 | `V20_10 | `V21_04 | `V21_10 | `V22_04 | `LTS | `Latest ]
@@ -163,7 +163,7 @@ let distros : t list = [
   `CentOS `V6; `CentOS `V7; `CentOS `V8; `CentOS `Latest;
   `Debian `V11; `Debian `V10; `Debian `V9; `Debian `V8; `Debian `V7;
   `Debian `Stable; `Debian `Testing; `Debian `Unstable;
-  `Fedora `V23; `Fedora `V24; `Fedora `V25; `Fedora `V26; `Fedora `V27; `Fedora `V28; `Fedora `V29; `Fedora `V30; `Fedora `V31; `Fedora `V32; `Fedora `V33; `Fedora `V34; `Fedora `Latest;
+  `Fedora `V23; `Fedora `V24; `Fedora `V25; `Fedora `V26; `Fedora `V27; `Fedora `V28; `Fedora `V29; `Fedora `V30; `Fedora `V31; `Fedora `V32; `Fedora `V33; `Fedora `V34; `Fedora `V35; `Fedora `Latest;
   `OracleLinux `V7; `OracleLinux `V8; `OracleLinux `Latest;
   `OpenSUSE `V42_1; `OpenSUSE `V42_2; `OpenSUSE `V42_3; `OpenSUSE `V15_0; `OpenSUSE `V15_1; `OpenSUSE `V15_2; `OpenSUSE `V15_3; `OpenSUSE `Latest;
   `Ubuntu `V12_04; `Ubuntu `V14_04; `Ubuntu `V15_04; `Ubuntu `V15_10;
@@ -235,8 +235,8 @@ let distro_status (d:t) : status = match d with
   | `Debian `Testing -> `Active `Tier3
   | `Debian `Unstable -> `Active `Tier3
   | `Fedora ( `V21 | `V22 | `V23 | `V24 | `V25 | `V26 | `V27 | `V28 | `V29 | `V30 | `V31 | `V32 | `V33) -> `Deprecated
-  | `Fedora `V34 -> `Active `Tier2
-  | `Fedora `Latest -> `Alias (`Fedora `V34)
+  | `Fedora (`V34 | `V35) -> `Active `Tier2
+  | `Fedora `Latest -> `Alias (`Fedora `V35)
   | `OracleLinux (`V7|`V8) -> `Active `Tier3
   | `OracleLinux `Latest -> `Alias (`OracleLinux `V8)
   | `OpenSUSE (`V42_1 | `V42_2 | `V42_3 | `V15_0 | `V15_1 | `V15_2) -> `Deprecated
@@ -287,7 +287,7 @@ let distro_arches ov (d:t) =
   | `Debian `V9, ov when OV.(compare Releases.v4_05_0 ov) = -1 -> [ `I386; `X86_64; `Aarch64; `Aarch32 ]
   | `Alpine (`V3_6 | `V3_7 | `V3_8 | `V3_9 | `V3_10 | `V3_11 | `V3_12 | `V3_13 | `V3_14 | `V3_15), ov when OV.(compare Releases.v4_05_0 ov) = -1 -> [ `X86_64; `Aarch64 ]
   | `Ubuntu (`V18_04|`V20_04|`V20_10|`V21_04|`V21_10|`V22_04), ov when OV.(compare Releases.v4_05_0 ov) = -1  -> [ `X86_64; `Aarch64; `Ppc64le ]
-  | `Fedora (`V33|`V34), ov when OV.(compare Releases.v4_08_0 ov) = -1  -> [ `X86_64; `Aarch64 ]
+  | `Fedora (`V33|`V34|`V35), ov when OV.(compare Releases.v4_08_0 ov) = -1  -> [ `X86_64; `Aarch64 ]
   (* 2021-04-19: should be 4.03 but there's a linking failure until 4.06. *)
   | `Windows (`Msvc, _), ov when OV.(compare Releases.v4_06_0 ov) = 1 -> []
   | _ -> [ `X86_64 ]
@@ -370,6 +370,7 @@ let builtin_ocaml_of_distro (d:t) : string option =
   |`Fedora `V32 -> Some "4.10.0"
   |`Fedora `V33 -> Some "4.11.1"
   |`Fedora `V34 -> Some "4.11.1"
+  |`Fedora `V35 -> Some "4.12.0"
   |`CentOS `V6 -> Some "3.11.2"
   |`CentOS `V7 -> Some "4.01.0"
   |`CentOS `V8 -> Some "4.07.0"
@@ -481,6 +482,7 @@ let tag_of_distro (d:t) = match d with
   |`Fedora `V32 -> "fedora-32"
   |`Fedora `V33 -> "fedora-33"
   |`Fedora `V34 -> "fedora-34"
+  |`Fedora `V35 -> "fedora-35"
   |`OracleLinux `V7 -> "oraclelinux-7"
   |`OracleLinux `V8 -> "oraclelinux-8"
   |`OracleLinux `Latest -> "oraclelinux"
@@ -563,6 +565,7 @@ let distro_of_tag x : t option =
   |"fedora-32" -> Some (`Fedora `V32)
   |"fedora-33" -> Some (`Fedora `V33)
   |"fedora-34" -> Some (`Fedora `V34)
+  |"fedora-35" -> Some (`Fedora `V35)
   |"fedora" -> Some (`Fedora `Latest)
   |"oraclelinux-7" -> Some (`OracleLinux `V7)
   |"oraclelinux-8" -> Some (`OracleLinux `V8)
@@ -643,6 +646,7 @@ let rec human_readable_string_of_distro (d:t) =
   |`Fedora `V32 -> "Fedora 32"
   |`Fedora `V33 -> "Fedora 33"
   |`Fedora `V34 -> "Fedora 34"
+  |`Fedora `V35 -> "Fedora 35"
   |`OracleLinux `V7 -> "OracleLinux 7"
   |`OracleLinux `V8 -> "OracleLinux 8"
   |`Alpine `V3_3 -> "Alpine 3.3"
@@ -809,6 +813,7 @@ let base_distro_tag ?win10_revision ?(arch=`X86_64) d =
         | `V32 -> "32"
         | `V33 -> "33"
         | `V34 -> "34"
+        | `V35 -> "35"
         | `Latest -> assert false
       in
       "fedora", tag